+++ src/components/index.ts
@@ -14,6 +14,8 @@ export * from "./draw/sprite";
 export * from "./draw/text";
 export * from "./draw/uvquad";
 export * from "./level/agent";
+export * from "./level/navigation";
+export * from "./level/patrol";
 export * from "./level/sentry";
 export * from "./level/tile";
 export * from "./misc/health";