+++ src/components/level/patrol.ts
@@ -1,60 +1,105 @@
 import { Vec2 } from "../../math";
 import type { Comp, GameObj, PosComp } from "../../types";
+import type { KEventController } from "../../utils";
 
 export interface PatrolComp extends Comp {
+    /*
+     * Path to follow. If null, doesn't move.
+     */
+    waypoints: Vec2[] | null;
+    /*
+     * Speed of the movement during patrol.
+     */
     patrolSpeed: number;
-    nextLocation: Vec2;
+    /*
+     * Current subgoal, if any.
+     */
+    nextLocation: Vec2 | null;
+    /*
+     * Only called when using "stop" and the path reaches its end.
+     */
+    onPatrolFinished(cb: (objects: GameObj[]) => void): KEventController;
 }
 
 type PatrolEndBehavior =
     /* Go directly bak to the start */
     | "loop"
-    | /* Traveling the reverse path back to the start */ "ping-pong"
-    | /* Stop */ "stop";
+    /* Traveling the reverse path back to the start */
+    | "ping-pong"
+    /* Stop */
+    | "stop";
 
 export interface PatrolCompOpt {
+    /*
+     * Path to follow. If null, starts suspended.
+     */
+    waypoints?: Vec2[];
+    /*
+     * Speed of the movement during patrol.
+     */
     speed?: number;
+    /*
+     * What to do after the last waypoint has been reached.
+     */
     endBehavior?: PatrolEndBehavior;
 }
 
 export function patrol(
-    waypoints: Vec2[],
     opts: PatrolCompOpt = {},
 ): PatrolComp {
+    let waypoints = opts.waypoints;
     let speed = opts.speed || 100; // Or throw error?
-    let endBehavior = opts.endBehavior || "ping-pong"; // Default is traveling the reverse path back to the start.
+    let endBehavior = opts.endBehavior || "stop"; // Default is stop.
     let index = 0;
+    let finished = waypoints != null;
     return {
         id: "patrol",
+        require: ["pos"],
         get patrolSpeed() {
             return speed;
         },
         set patrolSpeed(value) {
             speed = value;
         },
+        get waypoints() {
+            return waypoints;
+        },
+        set waypoints(value) {
+            waypoints = value;
+            index = 0;
+            finished = false;
+        },
         get nextLocation() {
-            return waypoints[index];
+            return waypoints ? waypoints[index] : null;
         },
         update(this: GameObj<PatrolComp | PosComp>) {
-            // TODO: When this object has agent, movement should happen through agent instead
-            this.moveTo(this.nextLocation, speed);
-            if (this.pos.sdist(this.nextLocation) < 4) {
+            const nextWaypoint = this.nextLocation
+            if (!waypoints || !nextWaypoint || finished) return;
+            this.moveTo(nextWaypoint, speed);
+            if (this.pos.sdist(nextWaypoint) < 9) {
                 switch (endBehavior) {
                     case "loop":
                         index = (index + 1) % waypoints.length;
                         break;
                     case "ping-pong":
                         index = index + 1;
-                        if (index = waypoints.length) {
+                        if (index == waypoints.length) {
                             waypoints.reverse();
                             index = 0;
                         }
                         break;
                     case "stop":
                         index = Math.min(index + 1, waypoints.length - 1);
+                        if (index == waypoints.length - 1) {
+                            finished = true;
+                            this.trigger("patrol-finished");
+                        }
                         break;
                 }
             }
+        },
+        onPatrolFinished(cb: (objects: GameObj[]) => void) {
+            return this.on("patrol-finished", cb);
         },
     };
 }